+++ Realm/RLMEmailPasswordAuth.mm
@@ -16,12 +16,14 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#import "RLMUsernamePasswordProviderClient.h"
+#import "RLMEmailPasswordAuth.h"
+#import "RLMProviderClient_Private.hpp"
+
 #import "RLMBSON_Private.hpp"
 #import "RLMApp_Private.hpp"
 #import "sync/app.hpp"
 
-@implementation RLMUsernamePasswordProviderClient
+@implementation RLMEmailPasswordAuth
 
 - (realm::app::App::UsernamePasswordProviderClient)client {
     return self.app._realmApp->provider_client<realm::app::App::UsernamePasswordProviderClient>();
@@ -29,52 +31,52 @@
 
 - (void)registerEmail:(NSString *)email
              password:(NSString *)password
-           completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+           completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.register_email(email.UTF8String, password.UTF8String, ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }
 
 - (void)confirmUser:(NSString *)token
             tokenId:(NSString *)tokenId
-         completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+         completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.confirm_user(token.UTF8String, tokenId.UTF8String, ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }
 
 - (void)resendConfirmationEmail:(NSString *)email
-                     completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+                     completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.resend_confirmation_email(email.UTF8String, ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }
 
 - (void)sendResetPasswordEmail:(NSString *)email
-                    completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+                    completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.send_reset_password_email(email.UTF8String, ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }
 
 - (void)resetPasswordTo:(NSString *)password
                   token:(NSString *)token
                 tokenId:(NSString *)tokenId
-             completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+             completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.reset_password(password.UTF8String, token.UTF8String, tokenId.UTF8String, ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }
 
 - (void)callResetPasswordFunction:(NSString *)email
                          password:(NSString *)password
                              args:(NSArray<id<RLMBSON>> *)args
-                       completion:(RLMUsernamePasswordProviderClientOptionalErrorBlock)completion {
+                       completion:(RLMEmailPasswordAuthOptionalErrorBlock)completion {
     self.client.call_reset_password_function(email.UTF8String,
                                              password.UTF8String,
                                              static_cast<realm::bson::BsonArray>(RLMConvertRLMBSONToBson(args)),
                                                                                  ^(realm::util::Optional<realm::app::AppError> error) {
-        [self.app handleResponse:error completion:completion];
+        [self handleResponse:error completion:completion];
     });
 }