+++ Realm/RLMPushClient.mm
@@ -18,7 +18,7 @@
 
 #import <Foundation/Foundation.h>
 #import "RLMPushClient_Private.hpp"
-#import "RLMSyncUser_Private.hpp"
+#import "RLMUser_Private.hpp"
 #import "RLMApp_Private.hpp"
 #import "sync/push_client.hpp"
 
@@ -34,7 +34,7 @@
     return nil;
 }
 
-- (void)registerDeviceWithToken:(NSString *)token user:(RLMSyncUser *)user completion:(RLMOptionalErrorBlock)completion {
+- (void)registerDeviceWithToken:(NSString *)token user:(RLMUser *)user completion:(RLMOptionalErrorBlock)completion {
     self->_pushClient->register_device(token.UTF8String, user._syncUser, ^(util::Optional<app::AppError> error) {
         if (error && error->error_code) {
             return completion(RLMAppErrorToNSError(*error));
@@ -44,7 +44,7 @@
 }
 
 
-- (void)deregisterDeviceForUser:(RLMSyncUser *)user completion:(RLMOptionalErrorBlock)completion {
+- (void)deregisterDeviceForUser:(RLMUser *)user completion:(RLMOptionalErrorBlock)completion {
     self->_pushClient->deregister_device(user._syncUser, ^(util::Optional<app::AppError> error) {
         if (error && error->error_code) {
             return completion(RLMAppErrorToNSError(*error));